refactor(front): add storage key constants
