+++ mylog/src/hooks/queries/useMarker.ts
@@ -1,7 +1,7 @@
 import {UseQueryOptions, useQuery} from '@tanstack/react-query';
 
 import {getMarkers} from '@/api/marker';
-import queryKeys from '@/constants/queryKeys';
+import {queryKeys} from '@/constants/keys';
 import type {Marker} from '@/types/domain';
 import type {ResponseError} from '@/types';
